refactor: remove appkit comps and export w3m components with multiple names
