--- packages/scaffold-ui/src/modal/appkit-network-button/styles.ts
@@ -1,8 +1,0 @@
-import { css } from 'lit'
-
-export default css`
-  :host {
-    display: block;
-    width: max-content;
-  }
-`
